Simplify PR: remove lp_rebalancer_2, simple_xemm_gateway; update lp_rebalancer
- Replace lp_rebalancer with lp_rebalancer_2 logic (uses executor limit prices
  for auto-close instead of timer-based rebalancing)
- Remove lp_rebalancer_2 directory (code now in lp_rebalancer)
- Remove simple_xemm_gateway.py script (not needed for this PR)
- Keep gateway_lp.py and gateway_swap.py deprecated stubs
