--- controllers/generic/lp_rebalancer_2/__init__.py
@@ -1,3 +1,0 @@
-from .lp_rebalancer_2 import LPRebalancer2, LPRebalancer2Config
-
-__all__ = ["LPRebalancer2", "LPRebalancer2Config"]
